fix(sidekick): move plugins.js to hub/scripts
