+++ hub/scripts/plugins.js
